Added Dockerfile and some cleanup
