+++ Website/src/app/api/auth/route.ts
@@ -1,23 +1,7 @@
 import {AuthenticationRequest, AuthenticationResponse} from "@/lib/api.website";
 import {cookies} from "next/headers";
 import {NextRequest, NextResponse} from "next/server";
-
-async function authenticate(username: string, password: string): Promise<string | undefined> {
-    console.log("Trying to authenticate with", username, password)
-    const auth_msg: AuthenticationRequest = { username: username, password: password };
-    const auth_resp_json = await fetch("http://localhost:8080/api/login/website", {
-        method: "POST", body: JSON.stringify(auth_msg), headers: {
-            "Content-Type": "application/json",
-            // 'Content-Type': 'application/x-www-form-urlencoded',
-        },
-    });
-    if ((await auth_resp_json).ok) {
-        const auth_resp: AuthenticationResponse = await auth_resp_json.json();
-        return auth_resp.token;
-    }
-    // return 'aaaaabbbbaaabbabaaa';
-    return;
-}
+import {authenticate} from "@/lib/data";
 
 // export async function GET(request: NextRequest) {
 //     return new NextResponse(null, { status: 405 })
